requirements.txt
Flask==3.0.3
requests==2.32.3
python-telegram-bot==20.7
